chore: refactor sidecar modules and docs
