+++ bolt-sidecar/src/chain_io/mod.rs
@@ -1,5 +1,6 @@
 /// Wrapper over the BoltManager contract
 pub mod manager;
+pub use manager::BoltManager;
 
 /// Utilities and functions used in the Bolt contracts
 pub mod utils;
