+++ bolt-sidecar/src/config/chain.rs
@@ -26,6 +26,7 @@ pub const APPLICATION_BUILDER_DOMAIN_MASK: [u8; 4] = [0, 0, 0, 1];
 /// The domain mask for signing commit-boost messages.
 pub const COMMIT_BOOST_DOMAIN_MASK: [u8; 4] = [109, 109, 111, 67];
 
+/// Default chain configuration for the sidecar.
 pub const DEFAULT_CHAIN_CONFIG: ChainConfig = ChainConfig {
     chain: Chain::Mainnet,
     commitment_deadline: DEFAULT_COMMITMENT_DEADLINE_IN_MILLIS,
@@ -87,6 +88,7 @@ impl Deref for ChainConfig {
 /// Supported chains for the sidecar
 #[derive(Debug, Clone, Copy, ValueEnum, Deserialize)]
 #[clap(rename_all = "kebab_case")]
+#[allow(missing_docs)]
 pub enum Chain {
     Mainnet,
     Holesky,
