+++ bolt-sidecar/src/state/mod.rs
@@ -1,7 +1,3 @@
-//! The `state` module is responsible for keeping a local copy of relevant state that is needed
-//! to simulate commitments against. It is updated on every block. It has both execution state and
-//! consensus state.
-
 use std::{
     pin::Pin,
     task::{Context, Poll},
@@ -11,6 +7,7 @@ use std::{
 use futures::{future::poll_fn, Future, FutureExt};
 use tokio::time::Sleep;
 
+/// Module to perform state validation.
 mod execution;
 pub use execution::{ExecutionState, ValidationError};
